Modularize E2EE replica tests
Separate round-trip, convergence, revision, snapshot, dirty-row, witness, and replica-apply coverage into focused suites while preserving production behavior.
