+++ crates/db-app/src/e2ee.rs
@@ -2062,3152 +2062,4 @@ fn push_json_bind(query: &mut QueryBuilder<Sqlite>, value: &Value) -> E2eeReplic
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use anlg_e2ee::RecoveryKey;
-
-    async fn test_db() -> anlg_db_core::Db {
-        let db = anlg_db_core::Db::connect_memory_plain().await.unwrap();
-        crate::prepare_schema(&db).await.unwrap();
-        db
-    }
-
-    fn keys(workspace_id: &str) -> HashMap<String, WorkspaceKey> {
-        let recovery =
-            RecoveryKey::parse("anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc")
-                .unwrap();
-        HashMap::from([(
-            workspace_id.to_string(),
-            recovery.workspace_key(workspace_id).unwrap(),
-        )])
-    }
-
-    async fn copy_replica(source: &SqlitePool, target: &SqlitePool) {
-        let records: Vec<(String, String, String)> =
-            sqlx::query_as("SELECT id, workspace_id, payload FROM e2ee_records")
-                .fetch_all(source)
-                .await
-                .unwrap();
-        for (id, workspace_id, payload) in records {
-            sqlx::query(
-                "INSERT INTO e2ee_records (id, workspace_id, payload) VALUES (?, ?, ?)
-                 ON CONFLICT(id) DO UPDATE SET payload = excluded.payload",
-            )
-            .bind(id)
-            .bind(workspace_id)
-            .bind(payload)
-            .execute(target)
-            .await
-            .unwrap();
-        }
-    }
-
-    #[tokio::test]
-    async fn encrypts_only_opaque_records() {
-        let db = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Secret planning')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let stats = encrypt_e2ee_replica_changes(db.pool(), &keys("workspace-a"))
-            .await
-            .unwrap();
-        let payloads: Vec<String> = sqlx::query_scalar("SELECT payload FROM e2ee_records")
-            .fetch_all(db.pool())
-            .await
-            .unwrap();
-
-        assert!(stats.encrypted_fields > 1);
-        assert!(!payloads.is_empty());
-        assert!(payloads.iter().all(|payload| {
-            !payload.contains("Secret planning")
-                && !payload.contains("session-1")
-                && !payload.contains("sessions")
-        }));
-    }
-
-    #[tokio::test]
-    async fn encrypts_and_reconstructs_attachment_metadata() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO session_attachments (
-               id, workspace_id, session_id, filename, relative_path,
-               content_type, size_bytes, sha256, source_type, source_id
-             ) VALUES (
-               'attachment-1', 'workspace-a', 'session-1', 'secret-diagram.png',
-               'attachments/secret-diagram.png', 'image/png', 42,
-               'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
-               'note_upload', 'secret-diagram.png'
-             )",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO attachment_local_state (
-               attachment_id, session_id, relative_path, availability
-             ) VALUES (
-               'attachment-1', 'session-1', 'local-only-secret.png', 'present'
-             )",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let payloads: Vec<String> = sqlx::query_scalar(
-            "SELECT payload FROM e2ee_records WHERE workspace_id = 'workspace-a'",
-        )
-        .fetch_all(source.pool())
-        .await
-        .unwrap();
-        assert!(!payloads.is_empty());
-        assert!(payloads.iter().all(|payload| {
-            !payload.contains("secret-diagram.png")
-                && !payload.contains("attachments/secret-diagram.png")
-                && !payload.contains("local-only-secret.png")
-                && !payload
-                    .contains("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa")
-        }));
-
-        let target = test_db().await;
-        copy_replica(source.pool(), target.pool()).await;
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let attachment: (String, String, String, i64, String) = sqlx::query_as(
-            "SELECT filename, relative_path, content_type, size_bytes, sha256
-             FROM session_attachments WHERE id = 'attachment-1'",
-        )
-        .fetch_one(target.pool())
-        .await
-        .unwrap();
-        assert_eq!(
-            attachment,
-            (
-                "secret-diagram.png".to_string(),
-                "attachments/secret-diagram.png".to_string(),
-                "image/png".to_string(),
-                42,
-                "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa".to_string(),
-            )
-        );
-        let local_state_count: i64 =
-            sqlx::query_scalar("SELECT COUNT(*) FROM attachment_local_state")
-                .fetch_one(target.pool())
-                .await
-                .unwrap();
-        assert_eq!(local_state_count, 0);
-    }
-
-    #[tokio::test]
-    async fn reconstructs_every_protected_table_and_applies_deletions() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        for (table, id) in [
-            ("action_items", "action-1"),
-            ("humans", "human-1"),
-            ("organizations", "organization-1"),
-            ("session_attachments", "attachment-1"),
-            ("session_documents", "document-1"),
-            ("session_participants", "participant-1"),
-            ("sessions", "session-1"),
-            ("transcripts", "transcript-1"),
-        ] {
-            let sql = format!("INSERT INTO {table} (id, workspace_id) VALUES (?, 'workspace-a')");
-            sqlx::query(sqlx::AssertSqlSafe(sql.as_str()))
-                .bind(id)
-                .execute(source.pool())
-                .await
-                .unwrap();
-        }
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let target = test_db().await;
-        copy_replica(source.pool(), target.pool()).await;
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        for table in E2EE_DOMAIN_TABLES {
-            let sql = format!("SELECT COUNT(*) FROM {table} WHERE workspace_id = 'workspace-a'");
-            let count: i64 = sqlx::query_scalar(sqlx::AssertSqlSafe(sql.as_str()))
-                .fetch_one(target.pool())
-                .await
-                .unwrap();
-            assert_eq!(count, 1, "{table} was not reconstructed");
-        }
-
-        sqlx::query("DELETE FROM sessions WHERE id = 'session-1'")
-            .execute(source.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        copy_replica(source.pool(), target.pool()).await;
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let sessions: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM sessions")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        assert_eq!(sessions, 0);
-    }
-
-    #[tokio::test]
-    async fn applies_remote_changes_and_preserves_concurrent_local_edits() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'First')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let target = test_db().await;
-        let records: Vec<(String, String, String)> =
-            sqlx::query_as("SELECT id, workspace_id, payload FROM e2ee_records")
-                .fetch_all(source.pool())
-                .await
-                .unwrap();
-        for (id, workspace_id, payload) in records {
-            sqlx::query("INSERT INTO e2ee_records (id, workspace_id, payload) VALUES (?, ?, ?)")
-                .bind(id)
-                .bind(workspace_id)
-                .bind(payload)
-                .execute(target.pool())
-                .await
-                .unwrap();
-        }
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        assert_eq!(title, "First");
-
-        sqlx::query("UPDATE sessions SET title = 'Remote' WHERE id = 'session-1'")
-            .execute(source.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let key = &workspace_keys["workspace-a"];
-        let title_record_id = key.blind_field_id("sessions", "session-1", "title");
-        let remote_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&title_record_id)
-                .fetch_one(source.pool())
-                .await
-                .unwrap();
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(remote_payload)
-            .bind(title_record_id)
-            .execute(target.pool())
-            .await
-            .unwrap();
-        sqlx::query("UPDATE sessions SET title = 'Local' WHERE id = 'session-1'")
-            .execute(target.pool())
-            .await
-            .unwrap();
-
-        let stats = apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-
-        assert_eq!(title, "Local");
-        assert_eq!(stats.skipped_local_changes, 1);
-    }
-
-    #[tokio::test]
-    async fn local_edits_advance_beyond_witnessed_remote_versions() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Base')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let remote_title = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "session-1",
-                "title",
-                "ffffffffffffffffffffffffffffffff",
-                8,
-                false,
-                json!("Remote"),
-            )
-            .unwrap();
-        let remote_tombstone = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "session-1",
-                ROW_MANIFEST_FIELD,
-                "ffffffffffffffffffffffffffffffff",
-                8,
-                true,
-                Value::Null,
-            )
-            .unwrap();
-        merge_e2ee_witness_events(
-            db.pool(),
-            key,
-            "workspace-a",
-            &[
-                E2eeWitnessEvent {
-                    sequence: 1,
-                    record_id: remote_title.record_id,
-                    workspace_id: "workspace-a".to_string(),
-                    payload_hash: anlg_e2ee::payload_hash(&remote_title.payload),
-                    payload: remote_title.payload,
-                },
-                E2eeWitnessEvent {
-                    sequence: 2,
-                    record_id: remote_tombstone.record_id,
-                    workspace_id: "workspace-a".to_string(),
-                    payload_hash: anlg_e2ee::payload_hash(&remote_tombstone.payload),
-                    payload: remote_tombstone.payload,
-                },
-            ],
-        )
-        .await
-        .unwrap();
-
-        sqlx::query("UPDATE sessions SET title = 'Local' WHERE id = 'session-1'")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let title_id = key.blind_field_id("sessions", "session-1", "title");
-        let manifest_id = key.blind_field_id("sessions", "session-1", ROW_MANIFEST_FIELD);
-        let title_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&title_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        let manifest_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&manifest_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        let title = key
-            .open_field("workspace-a", &title_id, &title_payload)
-            .unwrap();
-        let manifest = key
-            .open_field("workspace-a", &manifest_id, &manifest_payload)
-            .unwrap();
-
-        assert_eq!(title.revision, 9);
-        assert_eq!(title.value, json!("Local"));
-        assert_eq!(manifest.revision, 9);
-        assert!(!manifest.deleted);
-    }
-
-    #[tokio::test]
-    async fn encrypted_local_edit_advances_the_live_manifest() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Base')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        sqlx::query("UPDATE sessions SET title = 'Local' WHERE id = 'session-1'")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let manifest_id = key.blind_field_id("sessions", "session-1", ROW_MANIFEST_FIELD);
-        let local_manifest_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&manifest_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        let local_manifest = key
-            .open_field("workspace-a", &manifest_id, &local_manifest_payload)
-            .unwrap();
-        assert_eq!(local_manifest.revision, 2);
-
-        let remote_tombstone = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "session-1",
-                ROW_MANIFEST_FIELD,
-                "00000000000000000000000000000000",
-                2,
-                true,
-                Value::Null,
-            )
-            .unwrap();
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(remote_tombstone.payload)
-            .bind(&manifest_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        let stats = apply_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        let repaired_manifest: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(manifest_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-
-        assert_eq!(stats.rejected_rollbacks, 1);
-        assert_eq!(title, "Local");
-        assert_eq!(repaired_manifest, local_manifest_payload);
-    }
-
-    #[tokio::test]
-    async fn local_transcript_edit_rebases_above_witnessed_field_and_tombstone() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        let local_words = r#"[{"text":"preserve this local tail"}]"#;
-        sqlx::query(
-            "INSERT INTO transcripts (id, workspace_id, words_json)
-             VALUES ('transcript-1', 'workspace-a', '[{\"text\":\"base\"}]')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let words_id = key.blind_field_id("transcripts", "transcript-1", "words_json");
-        let manifest_id = key.blind_field_id("transcripts", "transcript-1", ROW_MANIFEST_FIELD);
-        let remote_words = key
-            .seal_field(
-                "workspace-a",
-                "transcripts",
-                "transcript-1",
-                "words_json",
-                "ffffffffffffffffffffffffffffffff",
-                7,
-                false,
-                json!(r#"[{"text":"remote"}]"#),
-            )
-            .unwrap();
-        let remote_tombstone = key
-            .seal_field(
-                "workspace-a",
-                "transcripts",
-                "transcript-1",
-                ROW_MANIFEST_FIELD,
-                "ffffffffffffffffffffffffffffffff",
-                8,
-                true,
-                Value::Null,
-            )
-            .unwrap();
-        merge_e2ee_witness_events(
-            db.pool(),
-            key,
-            "workspace-a",
-            &[
-                E2eeWitnessEvent {
-                    sequence: 1,
-                    record_id: words_id.clone(),
-                    workspace_id: "workspace-a".to_string(),
-                    payload_hash: anlg_e2ee::payload_hash(&remote_words.payload),
-                    payload: remote_words.payload.clone(),
-                },
-                E2eeWitnessEvent {
-                    sequence: 2,
-                    record_id: manifest_id.clone(),
-                    workspace_id: "workspace-a".to_string(),
-                    payload_hash: anlg_e2ee::payload_hash(&remote_tombstone.payload),
-                    payload: remote_tombstone.payload.clone(),
-                },
-            ],
-        )
-        .await
-        .unwrap();
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(&remote_words.payload)
-            .bind(&words_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(&remote_tombstone.payload)
-            .bind(&manifest_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        sqlx::query("UPDATE transcripts SET words_json = ? WHERE id = 'transcript-1'")
-            .bind(local_words)
-            .execute(db.pool())
-            .await
-            .unwrap();
-
-        let skipped = apply_e2ee_replica_changes_with_witness(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        assert!(skipped.skipped_local_changes > 0);
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let local_words_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&words_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        let local_manifest_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&manifest_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        let rebased_words = key
-            .open_field("workspace-a", &words_id, &local_words_payload)
-            .unwrap();
-        let rebased_manifest = key
-            .open_field("workspace-a", &manifest_id, &local_manifest_payload)
-            .unwrap();
-        assert_eq!(rebased_words.value, json!(local_words));
-        assert!(rebased_words.revision > 7);
-        assert!(!rebased_manifest.deleted);
-        assert!(rebased_manifest.revision > 8);
-
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(remote_tombstone.payload)
-            .bind(&manifest_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        apply_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let preserved: String =
-            sqlx::query_scalar("SELECT words_json FROM transcripts WHERE id = 'transcript-1'")
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert_eq!(preserved, local_words);
-    }
-
-    #[tokio::test]
-    async fn tombstoned_rows_can_be_recreated_with_retained_or_changed_fields() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        let target = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Same')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        copy_replica(source.pool(), target.pool()).await;
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        for title in ["Same", "Changed"] {
-            sqlx::query("DELETE FROM sessions WHERE id = 'session-1'")
-                .execute(source.pool())
-                .await
-                .unwrap();
-            encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-                .await
-                .unwrap();
-            copy_replica(source.pool(), target.pool()).await;
-            apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-                .await
-                .unwrap();
-            let deleted: i64 =
-                sqlx::query_scalar("SELECT COUNT(*) FROM sessions WHERE id = 'session-1'")
-                    .fetch_one(target.pool())
-                    .await
-                    .unwrap();
-            assert_eq!(deleted, 0);
-
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES ('session-1', 'workspace-a', 'user-a', ?)",
-            )
-            .bind(title)
-            .execute(source.pool())
-            .await
-            .unwrap();
-            encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-                .await
-                .unwrap();
-            copy_replica(source.pool(), target.pool()).await;
-            apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-                .await
-                .unwrap();
-            let materialized: String =
-                sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-                    .fetch_one(target.pool())
-                    .await
-                    .unwrap();
-            assert_eq!(materialized, title);
-        }
-    }
-
-    #[tokio::test]
-    async fn chunked_recreation_materializes_late_transcript_and_summary_fields() {
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        let source = test_db().await;
-        let target = test_db().await;
-        let words = r#"[{"text":"the complete transcript"}]"#;
-        let summary = r#"{"type":"doc","content":[{"type":"paragraph","text":"Full summary"}]}"#;
-        sqlx::query(
-            "INSERT INTO transcripts (id, workspace_id, words_json)
-             VALUES ('transcript-1', 'workspace-a', ?)",
-        )
-        .bind(words)
-        .execute(source.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO session_documents (id, workspace_id, kind, body)
-             VALUES ('summary-1', 'workspace-a', 'summary', ?)",
-        )
-        .bind(summary)
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        copy_replica(source.pool(), target.pool()).await;
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let words_id = key.blind_field_id("transcripts", "transcript-1", "words_json");
-        let body_id = key.blind_field_id("session_documents", "summary-1", "body");
-        let transcript_manifest_id =
-            key.blind_field_id("transcripts", "transcript-1", ROW_MANIFEST_FIELD);
-        let summary_manifest_id =
-            key.blind_field_id("session_documents", "summary-1", ROW_MANIFEST_FIELD);
-        let words_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&words_id)
-                .fetch_one(source.pool())
-                .await
-                .unwrap();
-        let body_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&body_id)
-                .fetch_one(source.pool())
-                .await
-                .unwrap();
-
-        sqlx::query("DELETE FROM transcripts WHERE id = 'transcript-1'")
-            .execute(source.pool())
-            .await
-            .unwrap();
-        sqlx::query("DELETE FROM session_documents WHERE id = 'summary-1'")
-            .execute(source.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        copy_replica(source.pool(), target.pool()).await;
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let transcript_tombstone_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&transcript_manifest_id)
-                .fetch_one(source.pool())
-                .await
-                .unwrap();
-        let summary_tombstone_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&summary_manifest_id)
-                .fetch_one(source.pool())
-                .await
-                .unwrap();
-        let transcript_tombstone = key
-            .open_field(
-                "workspace-a",
-                &transcript_manifest_id,
-                &transcript_tombstone_payload,
-            )
-            .unwrap();
-        let summary_tombstone = key
-            .open_field(
-                "workspace-a",
-                &summary_manifest_id,
-                &summary_tombstone_payload,
-            )
-            .unwrap();
-        let live_transcript = key
-            .seal_field(
-                "workspace-a",
-                "transcripts",
-                "transcript-1",
-                ROW_MANIFEST_FIELD,
-                "ffffffffffffffffffffffffffffffff",
-                transcript_tombstone.revision + 1,
-                false,
-                json!(true),
-            )
-            .unwrap();
-        let live_summary = key
-            .seal_field(
-                "workspace-a",
-                "session_documents",
-                "summary-1",
-                ROW_MANIFEST_FIELD,
-                "ffffffffffffffffffffffffffffffff",
-                summary_tombstone.revision + 1,
-                false,
-                json!(true),
-            )
-            .unwrap();
-
-        sqlx::query("DELETE FROM e2ee_records")
-            .execute(target.pool())
-            .await
-            .unwrap();
-        for (id, payload) in [
-            (&transcript_manifest_id, &live_transcript.payload),
-            (&summary_manifest_id, &live_summary.payload),
-        ] {
-            sqlx::query(
-                "INSERT INTO e2ee_records (id, workspace_id, payload)
-                 VALUES (?, 'workspace-a', ?)",
-            )
-            .bind(id)
-            .bind(payload)
-            .execute(target.pool())
-            .await
-            .unwrap();
-        }
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let defaults: (String, String) = (
-            sqlx::query_scalar("SELECT words_json FROM transcripts WHERE id = 'transcript-1'")
-                .fetch_one(target.pool())
-                .await
-                .unwrap(),
-            sqlx::query_scalar("SELECT body FROM session_documents WHERE id = 'summary-1'")
-                .fetch_one(target.pool())
-                .await
-                .unwrap(),
-        );
-        assert_eq!(defaults, ("[]".to_string(), String::new()));
-
-        for (id, payload) in [(&words_id, &words_payload), (&body_id, &body_payload)] {
-            sqlx::query(
-                "INSERT INTO e2ee_records (id, workspace_id, payload)
-                 VALUES (?, 'workspace-a', ?)",
-            )
-            .bind(id)
-            .bind(payload)
-            .execute(target.pool())
-            .await
-            .unwrap();
-        }
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let materialized: (String, String) = (
-            sqlx::query_scalar("SELECT words_json FROM transcripts WHERE id = 'transcript-1'")
-                .fetch_one(target.pool())
-                .await
-                .unwrap(),
-            sqlx::query_scalar("SELECT body FROM session_documents WHERE id = 'summary-1'")
-                .fetch_one(target.pool())
-                .await
-                .unwrap(),
-        );
-        assert_eq!(materialized, (words.to_string(), summary.to_string()));
-    }
-
-    #[tokio::test]
-    async fn recreation_advances_retained_fields_beyond_remote_tombstone_state() {
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        let local = test_db().await;
-        let remote = test_db().await;
-        for db in [&local, &remote] {
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES ('session-1', 'workspace-a', 'user-a', 'A')",
-            )
-            .execute(db.pool())
-            .await
-            .unwrap();
-            encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-                .await
-                .unwrap();
-        }
-
-        sqlx::query("UPDATE sessions SET title = 'B' WHERE id = 'session-1'")
-            .execute(remote.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(remote.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        sqlx::query("DELETE FROM sessions WHERE id = 'session-1'")
-            .execute(remote.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(remote.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let title_id = key.blind_field_id("sessions", "session-1", "title");
-        let manifest_id = key.blind_field_id("sessions", "session-1", ROW_MANIFEST_FIELD);
-        let remote_title_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&title_id)
-                .fetch_one(remote.pool())
-                .await
-                .unwrap();
-        let remote_manifest_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&manifest_id)
-                .fetch_one(remote.pool())
-                .await
-                .unwrap();
-        let remote_title = key
-            .open_field("workspace-a", &title_id, &remote_title_payload)
-            .unwrap();
-        let remote_manifest = key
-            .open_field("workspace-a", &manifest_id, &remote_manifest_payload)
-            .unwrap();
-
-        copy_replica(remote.pool(), local.pool()).await;
-        apply_e2ee_replica_changes(local.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        merge_e2ee_witness_events(
-            local.pool(),
-            key,
-            "workspace-a",
-            &[
-                E2eeWitnessEvent {
-                    sequence: 1,
-                    record_id: title_id.clone(),
-                    workspace_id: "workspace-a".to_string(),
-                    payload_hash: anlg_e2ee::payload_hash(&remote_title_payload),
-                    payload: remote_title_payload,
-                },
-                E2eeWitnessEvent {
-                    sequence: 2,
-                    record_id: manifest_id.clone(),
-                    workspace_id: "workspace-a".to_string(),
-                    payload_hash: anlg_e2ee::payload_hash(&remote_manifest_payload),
-                    payload: remote_manifest_payload,
-                },
-            ],
-        )
-        .await
-        .unwrap();
-        let deleted: i64 =
-            sqlx::query_scalar("SELECT COUNT(*) FROM sessions WHERE id = 'session-1'")
-                .fetch_one(local.pool())
-                .await
-                .unwrap();
-        assert_eq!(deleted, 0);
-
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'A')",
-        )
-        .execute(local.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(local.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let local_title_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&title_id)
-                .fetch_one(local.pool())
-                .await
-                .unwrap();
-        let local_manifest_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&manifest_id)
-                .fetch_one(local.pool())
-                .await
-                .unwrap();
-        let local_title = key
-            .open_field("workspace-a", &title_id, &local_title_payload)
-            .unwrap();
-        let local_manifest = key
-            .open_field("workspace-a", &manifest_id, &local_manifest_payload)
-            .unwrap();
-
-        assert_eq!(local_title.value, json!("A"));
-        assert!(local_title.revision > remote_title.revision);
-        assert!(!local_manifest.deleted);
-        assert!(local_manifest.revision > remote_manifest.revision);
-    }
-
-    #[tokio::test]
-    async fn rejects_authenticated_payloads_in_the_wrong_workspace() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        let sealed = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "session-1",
-                ROW_MANIFEST_FIELD,
-                "00000000000000000000000000000001",
-                1,
-                false,
-                json!(true),
-            )
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO e2ee_records (id, workspace_id, payload) VALUES (?, 'workspace-b', ?)",
-        )
-        .bind(sealed.record_id)
-        .bind(sealed.payload)
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let mut wrong_keys = workspace_keys;
-        wrong_keys.insert(
-            "workspace-b".to_string(),
-            RecoveryKey::parse("anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc")
-                .unwrap()
-                .workspace_key("workspace-b")
-                .unwrap(),
-        );
-        assert!(
-            apply_e2ee_replica_changes(db.pool(), &wrong_keys)
-                .await
-                .is_err()
-        );
-    }
-
-    #[tokio::test]
-    async fn rejects_and_repairs_replayed_older_field_revisions() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'First')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let record_id =
-            workspace_keys["workspace-a"].blind_field_id("sessions", "session-1", "title");
-        let old_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-
-        sqlx::query("UPDATE sessions SET title = 'Second' WHERE id = 'session-1'")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let current_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        apply_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(old_payload)
-            .bind(&record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-
-        let stats = apply_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        let repaired_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert_eq!(stats.rejected_rollbacks, 1);
-        assert_eq!(title, "Second");
-        assert_eq!(repaired_payload, current_payload);
-    }
-
-    #[tokio::test]
-    async fn equal_revision_payloads_converge_and_repair_the_replica() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Base')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let key = &workspace_keys["workspace-a"];
-        let first = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "session-1",
-                "title",
-                "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                2,
-                false,
-                json!("Device A"),
-            )
-            .unwrap();
-        let second = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "session-1",
-                "title",
-                "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
-                2,
-                false,
-                json!("Device B"),
-            )
-            .unwrap();
-        let (winner, winner_title, replay) = (second, "Device B", first);
-
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(&winner.payload)
-            .bind(&winner.record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        apply_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(&replay.payload)
-            .bind(&replay.record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        apply_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        let canonical_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&winner.record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert_eq!(title, winner_title);
-        assert_eq!(canonical_payload, winner.payload);
-
-        let third = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "session-1",
-                "title",
-                "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
-                3,
-                false,
-                json!("Clone A"),
-            )
-            .unwrap();
-        let fourth = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "session-1",
-                "title",
-                "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
-                3,
-                false,
-                json!("Clone B"),
-            )
-            .unwrap();
-        let (winner, winner_title, replay) =
-            if anlg_e2ee::payload_hash(&third.payload) > anlg_e2ee::payload_hash(&fourth.payload) {
-                (third, "Clone A", fourth)
-            } else {
-                (fourth, "Clone B", third)
-            };
-
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(&winner.payload)
-            .bind(&winner.record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        apply_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(&replay.payload)
-            .bind(&replay.record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        let stats = apply_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        let canonical_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&winner.record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert_eq!(stats.rejected_rollbacks, 1);
-        assert_eq!(title, winner_title);
-        assert_eq!(canonical_payload, winner.payload);
-    }
-
-    #[tokio::test]
-    async fn writer_ids_are_stable_per_device_and_unique_across_devices() {
-        let first = test_db().await;
-        sqlx::query("INSERT INTO sessions (id, workspace_id) VALUES ('first', 'workspace-a')")
-            .execute(first.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(first.pool(), &keys("workspace-a"))
-            .await
-            .unwrap();
-        let first_writer: String =
-            sqlx::query_scalar("SELECT writer_id FROM e2ee_local_device WHERE id = 'local'")
-                .fetch_one(first.pool())
-                .await
-                .unwrap();
-        encrypt_e2ee_replica_changes(first.pool(), &keys("workspace-a"))
-            .await
-            .unwrap();
-        let repeated_writer: String =
-            sqlx::query_scalar("SELECT writer_id FROM e2ee_local_device WHERE id = 'local'")
-                .fetch_one(first.pool())
-                .await
-                .unwrap();
-
-        let second = test_db().await;
-        sqlx::query("INSERT INTO sessions (id, workspace_id) VALUES ('second', 'workspace-a')")
-            .execute(second.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(second.pool(), &keys("workspace-a"))
-            .await
-            .unwrap();
-        let second_writer: String =
-            sqlx::query_scalar("SELECT writer_id FROM e2ee_local_device WHERE id = 'local'")
-                .fetch_one(second.pool())
-                .await
-                .unwrap();
-
-        assert_eq!(first_writer, repeated_writer);
-        assert_ne!(first_writer, second_writer);
-        assert_eq!(first_writer.len(), 32);
-        assert_eq!(second_writer.len(), 32);
-    }
-
-    #[tokio::test]
-    async fn fresh_device_does_not_materialize_an_unwitnessed_snapshot() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Archived value')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let fresh = test_db().await;
-        copy_replica(source.pool(), fresh.pool()).await;
-        let stats = apply_e2ee_replica_changes_with_witness(fresh.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let materialized: i64 =
-            sqlx::query_scalar("SELECT COUNT(*) FROM sessions WHERE id = 'session-1'")
-                .fetch_one(fresh.pool())
-                .await
-                .unwrap();
-        assert_eq!(materialized, 0);
-        assert!(stats.rejected_unwitnessed > 0);
-    }
-
-    #[tokio::test]
-    async fn witnessed_snapshot_materializes_and_repairs_the_replica() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Witnessed value')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let encrypted: Vec<(String, String)> = sqlx::query_as(
-            "SELECT id, payload FROM e2ee_records WHERE workspace_id = 'workspace-a' ORDER BY id",
-        )
-        .fetch_all(source.pool())
-        .await
-        .unwrap();
-        let events = encrypted
-            .iter()
-            .enumerate()
-            .map(|(index, (record_id, payload))| E2eeWitnessEvent {
-                sequence: u64::try_from(index + 1).unwrap(),
-                record_id: record_id.clone(),
-                workspace_id: "workspace-a".to_string(),
-                payload_hash: anlg_e2ee::payload_hash(payload),
-                payload: payload.clone(),
-            })
-            .collect::<Vec<_>>();
-
-        let fresh = test_db().await;
-        copy_replica(source.pool(), fresh.pool()).await;
-        merge_e2ee_witness_events(
-            fresh.pool(),
-            &workspace_keys["workspace-a"],
-            "workspace-a",
-            &events,
-        )
-        .await
-        .unwrap();
-        advance_e2ee_witness_cursor(fresh.pool(), "workspace-a", events.last().unwrap().sequence)
-            .await
-            .unwrap();
-        apply_e2ee_replica_changes_with_witness(fresh.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(fresh.pool())
-            .await
-            .unwrap();
-        assert_eq!(title, "Witnessed value");
-
-        sqlx::query("DELETE FROM e2ee_records")
-            .execute(fresh.pool())
-            .await
-            .unwrap();
-        apply_e2ee_replica_changes_with_witness(fresh.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let repaired: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*) FROM e2ee_records WHERE workspace_id = 'workspace-a'",
-        )
-        .fetch_one(fresh.pool())
-        .await
-        .unwrap();
-        assert_eq!(repaired, i64::try_from(events.len()).unwrap());
-    }
-
-    #[tokio::test]
-    async fn no_op_encryption_and_witness_repair_do_not_touch_replica_rows() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Unchanged')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let key = &workspace_keys["workspace-a"];
-        let title_record_id = key.blind_field_id("sessions", "session-1", "title");
-        let title_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&title_record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        sqlx::query("UPDATE e2ee_records SET updated_at = 'sentinel' WHERE id = ?")
-            .bind(&title_record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO e2ee_dirty_rows (workspace_id, table_name, row_id)
-             VALUES ('workspace-a', 'sessions', 'session-1')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let stats = encrypt_e2ee_replica_changes_bounded(db.pool(), &workspace_keys, 64)
-            .await
-            .unwrap();
-        let updated_at: String =
-            sqlx::query_scalar("SELECT updated_at FROM e2ee_records WHERE id = ?")
-                .bind(&title_record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert_eq!(stats.encrypted_fields, 0);
-        assert_eq!(updated_at, "sentinel");
-
-        merge_e2ee_witness_events(
-            db.pool(),
-            key,
-            "workspace-a",
-            &[E2eeWitnessEvent {
-                sequence: 1,
-                record_id: title_record_id.clone(),
-                workspace_id: "workspace-a".to_string(),
-                payload_hash: anlg_e2ee::payload_hash(&title_payload),
-                payload: title_payload,
-            }],
-        )
-        .await
-        .unwrap();
-        apply_received_e2ee_replica_changes_with_witness(db.pool(), &workspace_keys, true)
-            .await
-            .unwrap();
-        let repaired_updated_at: String =
-            sqlx::query_scalar("SELECT updated_at FROM e2ee_records WHERE id = ?")
-                .bind(title_record_id)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert_eq!(repaired_updated_at, "sentinel");
-    }
-
-    #[tokio::test]
-    async fn encrypts_only_rows_marked_dirty() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        for id in ["session-1", "session-2"] {
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES (?, 'workspace-a', 'user-a', ?)",
-            )
-            .bind(id)
-            .bind(id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-        }
-        sqlx::query(
-            "DELETE FROM e2ee_dirty_rows
-             WHERE workspace_id = 'workspace-a'
-               AND table_name = 'sessions'
-               AND row_id = 'session-2'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        encrypt_e2ee_replica_changes_bounded(db.pool(), &workspace_keys, 64)
-            .await
-            .unwrap();
-        let key = &workspace_keys["workspace-a"];
-        let first_manifest = key.blind_field_id("sessions", "session-1", ROW_MANIFEST_FIELD);
-        let second_manifest = key.blind_field_id("sessions", "session-2", ROW_MANIFEST_FIELD);
-        let first_exists: bool =
-            sqlx::query_scalar("SELECT EXISTS(SELECT 1 FROM e2ee_records WHERE id = ?)")
-                .bind(first_manifest)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        let second_exists: bool =
-            sqlx::query_scalar("SELECT EXISTS(SELECT 1 FROM e2ee_records WHERE id = ?)")
-                .bind(second_manifest)
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert!(first_exists);
-        assert!(!second_exists);
-    }
-
-    #[tokio::test]
-    async fn dirty_generation_preserves_a_concurrent_local_edit() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Before')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let writer_id = {
-            let mut transaction = db.pool().begin_with("BEGIN IMMEDIATE").await.unwrap();
-            let writer_id = load_or_create_writer_id(&mut transaction).await.unwrap();
-            transaction.commit().await.unwrap();
-            writer_id
-        };
-        let dirty = load_dirty_rows(db.pool(), &workspace_keys, 1)
-            .await
-            .unwrap()
-            .pop()
-            .unwrap();
-        let prepared =
-            prepare_dirty_row(db.pool(), &workspace_keys["workspace-a"], &writer_id, dirty)
-                .await
-                .unwrap();
-
-        sqlx::query("UPDATE sessions SET title = 'After' WHERE id = 'session-1'")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        let persisted = persist_prepared_dirty_row(db.pool(), prepared)
-            .await
-            .unwrap();
-        let generation: i64 = sqlx::query_scalar(
-            "SELECT generation FROM e2ee_dirty_rows
-             WHERE workspace_id = 'workspace-a'
-               AND table_name = 'sessions'
-               AND row_id = 'session-1'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        let replica_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_records")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        assert_eq!(persisted, 0);
-        assert_eq!(generation, 2);
-        assert_eq!(replica_count, 0);
-
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let key = &workspace_keys["workspace-a"];
-        let title_record_id = key.blind_field_id("sessions", "session-1", "title");
-        let payload: String = sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-            .bind(&title_record_id)
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        let field = key
-            .open_field("workspace-a", &title_record_id, &payload)
-            .unwrap();
-        assert_eq!(field.value, json!("After"));
-    }
-
-    #[tokio::test]
-    async fn capture_marker_inserted_after_prepare_prevents_transcript_persist() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        sqlx::query(
-            "INSERT INTO transcripts (id, workspace_id, session_id, words_json)
-             VALUES (
-               'transcript-1',
-               'workspace-a',
-               'session-1',
-               '[{\"text\":\"partial\"}]'
-             )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let writer_id = {
-            let mut transaction = db.pool().begin_with("BEGIN IMMEDIATE").await.unwrap();
-            let writer_id = load_or_create_writer_id(&mut transaction).await.unwrap();
-            transaction.commit().await.unwrap();
-            writer_id
-        };
-        let dirty = load_dirty_rows(db.pool(), &workspace_keys, 1)
-            .await
-            .unwrap()
-            .pop()
-            .unwrap();
-        let prepared =
-            prepare_dirty_row(db.pool(), &workspace_keys["workspace-a"], &writer_id, dirty)
-                .await
-                .unwrap();
-
-        sqlx::query(
-            "INSERT INTO app_settings (id, value_json)
-             VALUES ('capture_lifecycle_pending:session-1', ?)",
-        )
-        .bind(
-            json!({
-                "version": 1,
-                "phase": "capturing",
-                "sessionId": "session-1",
-                "transcriptId": "transcript-1",
-                "startedAt": 1_000,
-                "createdAt": "2026-07-24T00:00:00.000Z",
-                "audioOffsetMs": 0,
-                "preserveExistingTranscript": false,
-                "ownerUserId": "workspace-a",
-                "memo": ""
-            })
-            .to_string(),
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let persisted = persist_prepared_dirty_row_inner(db.pool(), prepared, true)
-            .await
-            .unwrap();
-        let dirty_count: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*) FROM e2ee_dirty_rows
-             WHERE table_name = 'transcripts' AND row_id = 'transcript-1'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        let local_state_count: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*) FROM e2ee_local_state
-             WHERE table_name = 'transcripts' AND row_id = 'transcript-1'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert_eq!(persisted, 0);
-        assert_eq!(dirty_count, 1);
-        assert_eq!(local_state_count, 0);
-
-        sqlx::query(
-            "UPDATE app_settings
-             SET value_json = json_set(value_json, '$.phase', 'finalizing')
-             WHERE id = 'capture_lifecycle_pending:session-1'",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes_deferring_active_captures(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let remaining: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        assert_eq!(remaining, 0);
-    }
-
-    #[tokio::test]
-    async fn cancelled_snapshot_finishes_current_row_and_releases_local_writes() {
-        let calibration_db = test_db().await;
-        let calibration_keys = keys("workspace-a");
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('calibration', 'workspace-a', 'user-a', 'Snapshot row')",
-        )
-        .execute(calibration_db.pool())
-        .await
-        .unwrap();
-        let calibration_checks = std::sync::atomic::AtomicUsize::new(0);
-        encrypt_e2ee_replica_changes_bounded_deferring_active_captures_cancellable(
-            calibration_db.pool(),
-            &calibration_keys,
-            E2EE_ENCRYPT_ROW_LIMIT,
-            || {
-                calibration_checks.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-                false
-            },
-        )
-        .await
-        .unwrap();
-        let cancel_after_first_persist = calibration_checks
-            .load(std::sync::atomic::Ordering::SeqCst)
-            .checked_sub(3)
-            .unwrap();
-
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        for index in 0..129 {
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES (?, 'workspace-a', 'user-a', 'Snapshot row')",
-            )
-            .bind(format!("session-{index:03}"))
-            .execute(db.pool())
-            .await
-            .unwrap();
-        }
-        let initial_dirty_rows: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        assert_eq!(initial_dirty_rows, 129);
-
-        let cancellation_checks = std::sync::atomic::AtomicUsize::new(0);
-        let stats = tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            encrypt_e2ee_replica_changes_bounded_deferring_active_captures_cancellable(
-                db.pool(),
-                &workspace_keys,
-                E2EE_ENCRYPT_ROW_LIMIT,
-                || {
-                    cancellation_checks.fetch_add(1, std::sync::atomic::Ordering::SeqCst)
-                        >= cancel_after_first_persist
-                },
-            ),
-        )
-        .await
-        .expect("snapshot cancellation exceeded the activity deadline")
-        .unwrap();
-
-        assert!(stats.remaining_replica_changes);
-        let remaining: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        assert_eq!(remaining, 128);
-
-        tokio::time::timeout(
-            std::time::Duration::from_millis(250),
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES ('after-cancel', 'workspace-a', 'user-a', 'Local write')",
-            )
-            .execute(db.pool()),
-        )
-        .await
-        .expect("cancelled snapshot kept the database busy")
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn large_dirty_queue_cancellation_stays_bounded_and_releases_local_writes() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        sqlx::query(
-            "WITH RECURSIVE rows(id) AS (
-                SELECT 1
-                UNION ALL
-                SELECT id + 1 FROM rows WHERE id < 100000
-            )
-            INSERT INTO e2ee_dirty_rows (workspace_id, table_name, row_id)
-            SELECT 'workspace-a', 'sessions', printf('missing-%06d', id) FROM rows",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let (page, remaining) =
-            load_dirty_rows_page(db.pool(), &workspace_keys, E2EE_ENCRYPT_ROW_LIMIT, true)
-                .await
-                .unwrap();
-        assert_eq!(page.len(), E2EE_ENCRYPT_ROW_LIMIT as usize);
-        assert!(remaining);
-
-        let cancellation_checks = std::sync::atomic::AtomicUsize::new(0);
-        let stats = tokio::time::timeout(
-            std::time::Duration::from_secs(1),
-            encrypt_e2ee_replica_changes_deferring_active_captures_cancellable(
-                db.pool(),
-                &workspace_keys,
-                || cancellation_checks.fetch_add(1, std::sync::atomic::Ordering::SeqCst) >= 1,
-            ),
-        )
-        .await
-        .expect("large dirty queue cancellation exceeded the activity deadline")
-        .unwrap();
-
-        assert!(stats.remaining_replica_changes);
-        let remaining: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        assert_eq!(remaining, 100_000);
-        tokio::time::timeout(
-            std::time::Duration::from_millis(250),
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES ('after-large-cancel', 'workspace-a', 'user-a', 'Local write')",
-            )
-            .execute(db.pool()),
-        )
-        .await
-        .expect("large dirty queue cancellation kept the database busy")
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn witness_change_preserves_a_prepared_dirty_row() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Local')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let writer_id = {
-            let mut transaction = db.pool().begin_with("BEGIN IMMEDIATE").await.unwrap();
-            let writer_id = load_or_create_writer_id(&mut transaction).await.unwrap();
-            transaction.commit().await.unwrap();
-            writer_id
-        };
-        let dirty = load_dirty_rows(db.pool(), &workspace_keys, 1)
-            .await
-            .unwrap()
-            .pop()
-            .unwrap();
-        let prepared = prepare_dirty_row(db.pool(), key, &writer_id, dirty)
-            .await
-            .unwrap();
-
-        let witnessed = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "session-1",
-                "title",
-                "ffffffffffffffffffffffffffffffff",
-                5,
-                false,
-                json!("Remote"),
-            )
-            .unwrap();
-        merge_e2ee_witness_events(
-            db.pool(),
-            key,
-            "workspace-a",
-            &[E2eeWitnessEvent {
-                sequence: 1,
-                record_id: witnessed.record_id,
-                workspace_id: "workspace-a".to_string(),
-                payload_hash: anlg_e2ee::payload_hash(&witnessed.payload),
-                payload: witnessed.payload,
-            }],
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(
-            persist_prepared_dirty_row(db.pool(), prepared)
-                .await
-                .unwrap(),
-            0
-        );
-        let dirty_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        let replica_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_records")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        assert_eq!(dirty_count, 1);
-        assert_eq!(replica_count, 0);
-
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let title_id = key.blind_field_id("sessions", "session-1", "title");
-        let payload: String = sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-            .bind(&title_id)
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        let field = key.open_field("workspace-a", &title_id, &payload).unwrap();
-        assert_eq!(field.revision, 6);
-        assert_eq!(field.value, json!("Local"));
-    }
-
-    #[tokio::test]
-    async fn bounded_encryption_makes_deterministic_progress() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        for index in 0..65 {
-            let id = format!("session-{index:03}");
-            sqlx::query("INSERT INTO sessions (id, workspace_id) VALUES (?, 'workspace-a')")
-                .bind(id)
-                .execute(db.pool())
-                .await
-                .unwrap();
-        }
-
-        encrypt_e2ee_replica_changes_bounded(db.pool(), &workspace_keys, 1_000)
-            .await
-            .unwrap();
-        let remaining: Vec<String> = sqlx::query_scalar(
-            "SELECT row_id FROM e2ee_dirty_rows
-             WHERE workspace_id = 'workspace-a'
-             ORDER BY row_id",
-        )
-        .fetch_all(db.pool())
-        .await
-        .unwrap();
-        assert_eq!(remaining, ["session-064"]);
-        assert!(
-            has_pending_e2ee_replica_changes(db.pool(), &workspace_keys)
-                .await
-                .unwrap()
-        );
-
-        encrypt_e2ee_replica_changes_bounded(db.pool(), &workspace_keys, 64)
-            .await
-            .unwrap();
-        let remaining_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        assert_eq!(remaining_count, 0);
-        assert!(
-            !has_pending_e2ee_replica_changes(db.pool(), &workspace_keys)
-                .await
-                .unwrap()
-        );
-    }
-
-    #[tokio::test]
-    async fn active_capture_transcripts_do_not_starve_unrelated_rows() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        for index in 0..65 {
-            let transcript_id = format!("protected-{index:03}");
-            let session_id = format!("session-{index:03}");
-            sqlx::query(
-                "INSERT INTO transcripts (id, workspace_id, words_json)
-                 VALUES (?, 'workspace-a', '[{\"text\":\"partial\"}]')",
-            )
-            .bind(&transcript_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-            sqlx::query("INSERT INTO app_settings (id, value_json) VALUES (?, ?)")
-                .bind(format!("capture_lifecycle_pending:{session_id}"))
-                .bind(
-                    json!({
-                        "version": 1,
-                        "phase": "capturing",
-                        "sessionId": session_id,
-                        "transcriptId": transcript_id,
-                        "startedAt": 1_000,
-                        "createdAt": "2026-07-24T00:00:00.000Z",
-                        "audioOffsetMs": 0,
-                        "preserveExistingTranscript": false,
-                        "ownerUserId": "workspace-a",
-                        "memo": ""
-                    })
-                    .to_string(),
-                )
-                .execute(db.pool())
-                .await
-                .unwrap();
-        }
-        sqlx::query(
-            "INSERT INTO transcripts (id, workspace_id, words_json)
-             VALUES ('unrelated', 'workspace-a', '[{\"text\":\"complete\"}]')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        encrypt_e2ee_replica_changes_bounded_deferring_active_captures(
-            db.pool(),
-            &workspace_keys,
-            64,
-        )
-        .await
-        .unwrap();
-
-        let remaining_protected: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*) FROM e2ee_dirty_rows
-             WHERE table_name = 'transcripts' AND row_id GLOB 'protected-*'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        let unrelated_dirty: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*) FROM e2ee_dirty_rows
-             WHERE table_name = 'transcripts' AND row_id = 'unrelated'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        let unrelated_records: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*) FROM e2ee_local_state
-             WHERE table_name = 'transcripts' AND row_id = 'unrelated'",
-        )
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert_eq!(remaining_protected, 65);
-        assert_eq!(unrelated_dirty, 0);
-        assert!(unrelated_records > 0);
-        assert!(
-            !has_pending_e2ee_replica_changes_deferring_active_captures(
-                db.pool(),
-                &workspace_keys,
-            )
-            .await
-            .unwrap()
-        );
-        assert!(
-            has_pending_e2ee_replica_changes(db.pool(), &workspace_keys)
-                .await
-                .unwrap()
-        );
-
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let remaining: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(db.pool())
-            .await
-            .unwrap();
-        assert_eq!(remaining, 0);
-    }
-
-    #[tokio::test]
-    async fn pending_witness_uploads_respect_record_and_byte_limits() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Bounded witness uploads')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let key = &workspace_keys["workspace-a"];
-
-        let first = pending_e2ee_witness_uploads(db.pool(), "workspace-a", key, 1, usize::MAX)
-            .await
-            .unwrap();
-        assert_eq!(first.len(), 1);
-        let first_bytes = first[0]
-            .payload
-            .len()
-            .saturating_add(first[0].record_id.len())
-            .saturating_add(first[0].payload_hash.len())
-            .saturating_add(256);
-        let byte_bounded =
-            pending_e2ee_witness_uploads(db.pool(), "workspace-a", key, 128, first_bytes)
-                .await
-                .unwrap();
-        assert_eq!(byte_bounded, first);
-        assert!(matches!(
-            pending_e2ee_witness_uploads(db.pool(), "workspace-a", key, 128, first_bytes - 1,)
-                .await,
-            Err(E2eeReplicaError::WitnessUploadTooLarge)
-        ));
-
-        acknowledge_e2ee_witness_uploads(db.pool(), key, &byte_bounded)
-            .await
-            .unwrap();
-        let next = pending_e2ee_witness_uploads(db.pool(), "workspace-a", key, 1, usize::MAX)
-            .await
-            .unwrap();
-        assert_eq!(next.len(), 1);
-        assert_ne!(next[0].record_id, first[0].record_id);
-    }
-
-    #[tokio::test]
-    async fn pending_witness_uploads_never_scan_all_local_state() {
-        let db = test_db().await;
-        let explain = format!("EXPLAIN QUERY PLAN {PENDING_E2EE_WITNESS_UPLOADS_SQL}");
-        let plan: Vec<(i64, i64, i64, String)> =
-            sqlx::query_as(sqlx::AssertSqlSafe(explain.as_str()))
-                .bind("workspace-a")
-                .bind(16_i64)
-                .fetch_all(db.pool())
-                .await
-                .unwrap();
-        let details = plan
-            .into_iter()
-            .map(|(_, _, _, detail)| detail)
-            .collect::<Vec<_>>();
-
-        assert!(details.iter().any(|detail| {
-            detail.contains(
-                "SEARCH pending USING COVERING INDEX idx_e2ee_witness_pending_workspace_record",
-            )
-        }));
-        assert!(!details.iter().any(|detail| detail.contains("SCAN local")));
-    }
-
-    #[tokio::test]
-    async fn local_edits_enqueue_and_acknowledgements_drain_witness_uploads() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Before')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let initial = pending_e2ee_witness_uploads(db.pool(), "workspace-a", key, 128, usize::MAX)
-            .await
-            .unwrap();
-        assert!(!initial.is_empty());
-        acknowledge_e2ee_witness_uploads(db.pool(), key, &initial)
-            .await
-            .unwrap();
-        let pending_after_ack: i64 =
-            sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_witness_pending")
-                .fetch_one(db.pool())
-                .await
-                .unwrap();
-        assert_eq!(pending_after_ack, 0);
-
-        sqlx::query("UPDATE sessions SET title = 'After' WHERE id = 'session-1'")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let title_record_id = key.blind_field_id("sessions", "session-1", "title");
-        let title_pending: bool = sqlx::query_scalar(
-            "SELECT EXISTS(
-               SELECT 1 FROM e2ee_witness_pending WHERE record_id = ?
-             )",
-        )
-        .bind(title_record_id)
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert!(title_pending);
-    }
-
-    #[tokio::test]
-    async fn stale_witness_ack_keeps_a_newer_local_upload_pending() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'First')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let initial = pending_e2ee_witness_uploads(db.pool(), "workspace-a", key, 128, usize::MAX)
-            .await
-            .unwrap();
-        let title_record_id = key.blind_field_id("sessions", "session-1", "title");
-        let stale_title = initial
-            .iter()
-            .find(|upload| upload.record_id == title_record_id)
-            .unwrap()
-            .clone();
-        acknowledge_e2ee_witness_uploads(db.pool(), key, &initial)
-            .await
-            .unwrap();
-
-        sqlx::query("UPDATE sessions SET title = 'Second' WHERE id = 'session-1'")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        acknowledge_e2ee_witness_uploads(db.pool(), key, std::slice::from_ref(&stale_title))
-            .await
-            .unwrap();
-
-        let pending = pending_e2ee_witness_uploads(db.pool(), "workspace-a", key, 128, usize::MAX)
-            .await
-            .unwrap();
-        let current_title = pending
-            .iter()
-            .find(|upload| upload.record_id == title_record_id)
-            .unwrap();
-        assert!(current_title.revision > stale_title.revision);
-        assert_ne!(current_title.payload, stale_title.payload);
-
-        acknowledge_e2ee_witness_uploads(db.pool(), key, std::slice::from_ref(current_title))
-            .await
-            .unwrap();
-        let title_pending: bool = sqlx::query_scalar(
-            "SELECT EXISTS(
-               SELECT 1 FROM e2ee_witness_pending WHERE record_id = ?
-             )",
-        )
-        .bind(title_record_id)
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert!(!title_pending);
-    }
-
-    #[tokio::test]
-    async fn equal_remote_state_does_not_enqueue_a_witness_upload() {
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Remote')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let uploads =
-            pending_e2ee_witness_uploads(source.pool(), "workspace-a", key, 128, usize::MAX)
-                .await
-                .unwrap();
-        let events = uploads
-            .iter()
-            .enumerate()
-            .map(|(index, upload)| E2eeWitnessEvent {
-                sequence: u64::try_from(index + 1).unwrap(),
-                record_id: upload.record_id.clone(),
-                workspace_id: upload.workspace_id.clone(),
-                payload_hash: upload.payload_hash.clone(),
-                payload: upload.payload.clone(),
-            })
-            .collect::<Vec<_>>();
-
-        let target = test_db().await;
-        copy_replica(source.pool(), target.pool()).await;
-        merge_e2ee_witness_events(target.pool(), key, "workspace-a", &events)
-            .await
-            .unwrap();
-        let stats = apply_e2ee_replica_changes_with_witness(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        let pending: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_witness_pending")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        assert!(stats.applied_fields > 0);
-        assert_eq!(title, "Remote");
-        assert_eq!(pending, 0);
-    }
-
-    #[tokio::test]
-    async fn deleting_local_state_cascades_to_the_witness_queue() {
-        let db = test_db().await;
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Pending')",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(db.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let title_record_id = key.blind_field_id("sessions", "session-1", "title");
-        let pending_before: bool = sqlx::query_scalar(
-            "SELECT EXISTS(
-               SELECT 1 FROM e2ee_witness_pending WHERE record_id = ?
-             )",
-        )
-        .bind(&title_record_id)
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert!(pending_before);
-
-        sqlx::query("DELETE FROM e2ee_local_state WHERE record_id = ?")
-            .bind(&title_record_id)
-            .execute(db.pool())
-            .await
-            .unwrap();
-
-        let pending_after: bool = sqlx::query_scalar(
-            "SELECT EXISTS(
-               SELECT 1 FROM e2ee_witness_pending WHERE record_id = ?
-             )",
-        )
-        .bind(title_record_id)
-        .fetch_one(db.pool())
-        .await
-        .unwrap();
-        assert!(!pending_after);
-    }
-
-    #[tokio::test]
-    async fn remote_apply_guard_preserves_an_existing_local_dirty_marker() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title, status)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'First', 'active')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let target = test_db().await;
-        copy_replica(source.pool(), target.pool()).await;
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        sqlx::query("UPDATE sessions SET title = 'Local' WHERE id = 'session-1'")
-            .execute(target.pool())
-            .await
-            .unwrap();
-        let generation_before: i64 = sqlx::query_scalar(
-            "SELECT generation FROM e2ee_dirty_rows
-             WHERE workspace_id = 'workspace-a'
-               AND table_name = 'sessions'
-               AND row_id = 'session-1'",
-        )
-        .fetch_one(target.pool())
-        .await
-        .unwrap();
-
-        sqlx::query("UPDATE sessions SET status = 'archived' WHERE id = 'session-1'")
-            .execute(source.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let key = &workspace_keys["workspace-a"];
-        let status_record_id = key.blind_field_id("sessions", "session-1", "status");
-        let status_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&status_record_id)
-                .fetch_one(source.pool())
-                .await
-                .unwrap();
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(status_payload)
-            .bind(status_record_id)
-            .execute(target.pool())
-            .await
-            .unwrap();
-
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let row: (String, String) =
-            sqlx::query_as("SELECT title, status FROM sessions WHERE id = 'session-1'")
-                .fetch_one(target.pool())
-                .await
-                .unwrap();
-        let generation_after: i64 = sqlx::query_scalar(
-            "SELECT generation FROM e2ee_dirty_rows
-             WHERE workspace_id = 'workspace-a'
-               AND table_name = 'sessions'
-               AND row_id = 'session-1'",
-        )
-        .fetch_one(target.pool())
-        .await
-        .unwrap();
-        assert_eq!(row, ("Local".to_string(), "archived".to_string()));
-        assert_eq!(generation_after, generation_before);
-
-        let encrypted = encrypt_e2ee_replica_changes_bounded(target.pool(), &workspace_keys, 64)
-            .await
-            .unwrap();
-        let title_record_id = key.blind_field_id("sessions", "session-1", "title");
-        let title_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&title_record_id)
-                .fetch_one(target.pool())
-                .await
-                .unwrap();
-        let title_field = key
-            .open_field("workspace-a", &title_record_id, &title_payload)
-            .unwrap();
-        let remaining_dirty: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        assert!(encrypted.encrypted_fields > 0);
-        assert_eq!(title_field.value, json!("Local"));
-        assert_eq!(remaining_dirty, 0);
-    }
-
-    #[tokio::test]
-    async fn field_chunk_before_manifest_waits_then_applies_without_echo() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Chunked')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let key = &workspace_keys["workspace-a"];
-        let title_id = key.blind_field_id("sessions", "session-1", "title");
-        let manifest_id = key.blind_field_id("sessions", "session-1", ROW_MANIFEST_FIELD);
-        let title_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&title_id)
-                .fetch_one(source.pool())
-                .await
-                .unwrap();
-        let manifest_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&manifest_id)
-                .fetch_one(source.pool())
-                .await
-                .unwrap();
-
-        let target = test_db().await;
-        sqlx::query(
-            "INSERT INTO e2ee_records (id, workspace_id, payload)
-             VALUES (?, 'workspace-a', ?)",
-        )
-        .bind(&title_id)
-        .bind(title_payload)
-        .execute(target.pool())
-        .await
-        .unwrap();
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let before_manifest: i64 =
-            sqlx::query_scalar("SELECT COUNT(*) FROM sessions WHERE id = 'session-1'")
-                .fetch_one(target.pool())
-                .await
-                .unwrap();
-        assert_eq!(before_manifest, 0);
-
-        sqlx::query(
-            "INSERT INTO e2ee_records (id, workspace_id, payload)
-             VALUES (?, 'workspace-a', ?)",
-        )
-        .bind(manifest_id)
-        .bind(manifest_payload)
-        .execute(target.pool())
-        .await
-        .unwrap();
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        let dirty_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_dirty_rows")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        let outbound = encrypt_e2ee_replica_changes_bounded(target.pool(), &workspace_keys, 64)
-            .await
-            .unwrap();
-        assert_eq!(title, "Chunked");
-        assert_eq!(dirty_count, 0);
-        assert_eq!(outbound.encrypted_fields, 0);
-    }
-
-    #[tokio::test]
-    async fn witness_repairs_missing_records_only_after_snapshot_completion() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Witness recovery')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let encrypted: Vec<(String, String)> = sqlx::query_as(
-            "SELECT id, payload FROM e2ee_records
-             WHERE workspace_id = 'workspace-a'
-             ORDER BY id",
-        )
-        .fetch_all(source.pool())
-        .await
-        .unwrap();
-        let events = encrypted
-            .into_iter()
-            .enumerate()
-            .map(|(index, (record_id, payload))| E2eeWitnessEvent {
-                sequence: u64::try_from(index + 1).unwrap(),
-                record_id,
-                workspace_id: "workspace-a".to_string(),
-                payload_hash: anlg_e2ee::payload_hash(&payload),
-                payload,
-            })
-            .collect::<Vec<_>>();
-
-        let target = test_db().await;
-        merge_e2ee_witness_events(
-            target.pool(),
-            &workspace_keys["workspace-a"],
-            "workspace-a",
-            &events,
-        )
-        .await
-        .unwrap();
-        apply_received_e2ee_replica_changes_with_witness(target.pool(), &workspace_keys, false)
-            .await
-            .unwrap();
-        let partial_replica_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_records")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        let partial_session_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM sessions")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        assert_eq!(partial_replica_count, 0);
-        assert_eq!(partial_session_count, 0);
-
-        apply_received_e2ee_replica_changes_with_witness(target.pool(), &workspace_keys, true)
-            .await
-            .unwrap();
-        let title: String = sqlx::query_scalar("SELECT title FROM sessions WHERE id = 'session-1'")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        assert_eq!(title, "Witness recovery");
-    }
-
-    #[tokio::test]
-    async fn witness_repairs_respect_record_and_byte_limits() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Bounded witness recovery')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let encrypted: Vec<(String, String)> = sqlx::query_as(
-            "SELECT id, payload FROM e2ee_records
-             WHERE workspace_id = 'workspace-a'
-             ORDER BY id",
-        )
-        .fetch_all(source.pool())
-        .await
-        .unwrap();
-        assert!(encrypted.len() > 2);
-        let events = encrypted
-            .iter()
-            .enumerate()
-            .map(|(index, (record_id, payload))| E2eeWitnessEvent {
-                sequence: u64::try_from(index + 1).unwrap(),
-                record_id: record_id.clone(),
-                workspace_id: "workspace-a".to_string(),
-                payload_hash: anlg_e2ee::payload_hash(payload),
-                payload: payload.clone(),
-            })
-            .collect::<Vec<_>>();
-
-        let target = test_db().await;
-        merge_e2ee_witness_events(
-            target.pool(),
-            &workspace_keys["workspace-a"],
-            "workspace-a",
-            &events,
-        )
-        .await
-        .unwrap();
-
-        let error =
-            repair_e2ee_replica_from_witness_bounded(target.pool(), &workspace_keys, true, 2, 1)
-                .await
-                .unwrap_err();
-        assert!(matches!(error, E2eeReplicaError::WitnessRepairTooLarge));
-        let empty_replica: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_records")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        assert_eq!(empty_replica, 0);
-
-        let first = repair_e2ee_replica_from_witness_bounded(
-            target.pool(),
-            &workspace_keys,
-            true,
-            2,
-            usize::MAX,
-        )
-        .await
-        .unwrap();
-        assert_eq!(first.repaired_records, 2);
-        assert!(first.remaining);
-
-        let mut remaining = first.remaining;
-        while remaining {
-            let outcome = repair_e2ee_replica_from_witness_bounded(
-                target.pool(),
-                &workspace_keys,
-                true,
-                2,
-                usize::MAX,
-            )
-            .await
-            .unwrap();
-            assert!(outcome.repaired_records <= 2);
-            remaining = outcome.remaining;
-        }
-
-        let repaired: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_records")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        assert_eq!(repaired, i64::try_from(events.len()).unwrap());
-        assert!(
-            !has_pending_e2ee_witness_repairs(target.pool(), &workspace_keys, true)
-                .await
-                .unwrap()
-        );
-    }
-
-    #[tokio::test]
-    async fn completed_snapshots_repair_large_witness_sets_in_bounded_cycles() {
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        let events = (0..130)
-            .map(|index| {
-                let row_id = format!("session-{index:03}");
-                let sealed = key
-                    .seal_field(
-                        "workspace-a",
-                        "sessions",
-                        &row_id,
-                        ROW_MANIFEST_FIELD,
-                        "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                        1,
-                        false,
-                        json!(true),
-                    )
-                    .unwrap();
-                E2eeWitnessEvent {
-                    sequence: u64::try_from(index + 1).unwrap(),
-                    record_id: sealed.record_id,
-                    workspace_id: "workspace-a".to_string(),
-                    payload_hash: anlg_e2ee::payload_hash(&sealed.payload),
-                    payload: sealed.payload,
-                }
-            })
-            .collect::<Vec<_>>();
-
-        let target = test_db().await;
-        merge_e2ee_witness_events(target.pool(), key, "workspace-a", &events)
-            .await
-            .unwrap();
-        let max_record_bytes: i64 = sqlx::query_scalar(
-            "SELECT MAX(
-               LENGTH(CAST(workspace_id AS BLOB))
-                 + LENGTH(CAST(record_id AS BLOB))
-                 + LENGTH(CAST(payload AS BLOB))
-                 + 256
-             )
-             FROM e2ee_witness_records
-             WHERE workspace_id = 'workspace-a'",
-        )
-        .fetch_one(target.pool())
-        .await
-        .unwrap();
-        let max_repair_bytes = usize::try_from(max_record_bytes).unwrap() * 7;
-        let mut cycles = 0;
-
-        loop {
-            let before: (i64, i64) = sqlx::query_as(
-                "SELECT
-                   COUNT(*),
-                   COALESCE(SUM(
-                     LENGTH(CAST(workspace_id AS BLOB))
-                       + LENGTH(CAST(id AS BLOB))
-                       + LENGTH(CAST(payload AS BLOB))
-                       + 256
-                   ), 0)
-                 FROM e2ee_records",
-            )
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-            let stats = apply_received_e2ee_replica_changes_with_witness_bounded(
-                target.pool(),
-                &workspace_keys,
-                true,
-                64,
-                max_repair_bytes,
-                &|| false,
-            )
-            .await
-            .unwrap();
-            let after: (i64, i64) = sqlx::query_as(
-                "SELECT
-                   COUNT(*),
-                   COALESCE(SUM(
-                     LENGTH(CAST(workspace_id AS BLOB))
-                       + LENGTH(CAST(id AS BLOB))
-                       + LENGTH(CAST(payload AS BLOB))
-                       + 256
-                   ), 0)
-                 FROM e2ee_records",
-            )
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-            let repaired_records = after.0 - before.0;
-            let repaired_bytes = after.1 - before.1;
-            assert!(repaired_records > 0);
-            assert!(repaired_records <= 64);
-            assert!(usize::try_from(repaired_bytes).unwrap() <= max_repair_bytes);
-            cycles += 1;
-
-            let pending = has_pending_e2ee_replica_changes(target.pool(), &workspace_keys)
-                .await
-                .unwrap();
-            assert_eq!(stats.remaining_replica_changes, pending);
-            if !pending {
-                break;
-            }
-            assert!(cycles < 130);
-        }
-
-        assert!(cycles > 1);
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM e2ee_records")
-                .fetch_one(target.pool())
-                .await
-                .unwrap(),
-            i64::try_from(events.len()).unwrap()
-        );
-    }
-
-    #[tokio::test]
-    async fn received_replica_rows_apply_in_bounded_cycles() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        for index in 0..20 {
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES (?, 'workspace-a', 'user-a', ?)",
-            )
-            .bind(format!("remote-session-{index:02}"))
-            .bind(format!("Remote {index:02}"))
-            .execute(source.pool())
-            .await
-            .unwrap();
-        }
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let target = test_db().await;
-        copy_replica(source.pool(), target.pool()).await;
-        let mut stats = apply_e2ee_replica_changes_inner(
-            target.pool(),
-            &workspace_keys,
-            false,
-            3,
-            usize::MAX,
-            &|| false,
-        )
-        .await
-        .unwrap();
-        assert!(stats.remaining_replica_changes);
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>(
-                "SELECT COUNT(*) FROM sessions WHERE id LIKE 'remote-session-%'",
-            )
-            .fetch_one(target.pool())
-            .await
-            .unwrap(),
-            3
-        );
-
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('local-session', 'workspace-a', 'user-a', 'Local')",
-        )
-        .execute(target.pool())
-        .await
-        .unwrap();
-        let mut cycles = 1;
-        while stats.remaining_replica_changes {
-            stats = apply_e2ee_replica_changes_inner(
-                target.pool(),
-                &workspace_keys,
-                false,
-                3,
-                usize::MAX,
-                &|| false,
-            )
-            .await
-            .unwrap();
-            sqlx::query("UPDATE sessions SET title = title WHERE id = 'local-session'")
-                .execute(target.pool())
-                .await
-                .unwrap();
-            cycles += 1;
-            assert!(cycles < 20);
-        }
-
-        assert!(cycles > 1);
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>(
-                "SELECT COUNT(*) FROM sessions WHERE id LIKE 'remote-session-%'",
-            )
-            .fetch_one(target.pool())
-            .await
-            .unwrap(),
-            20
-        );
-        assert_eq!(
-            sqlx::query_scalar::<_, String>(
-                "SELECT title FROM sessions WHERE id = 'local-session'",
-            )
-            .fetch_one(target.pool())
-            .await
-            .unwrap(),
-            "Local"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancelled_received_replica_preflight_is_bounded_and_releases_local_writes() {
-        use std::sync::atomic::{AtomicUsize, Ordering as AtomicOrdering};
-
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        let db = test_db().await;
-        let mut transaction = db.pool().begin().await.unwrap();
-        for index in 0..512 {
-            let row_id = format!("remote-session-{index:04}");
-            let sealed = key
-                .seal_field(
-                    "workspace-a",
-                    "sessions",
-                    &row_id,
-                    ROW_MANIFEST_FIELD,
-                    "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                    1,
-                    false,
-                    json!(true),
-                )
-                .unwrap();
-            let payload_hash = anlg_e2ee::payload_hash(&sealed.payload);
-            sqlx::query(
-                "INSERT INTO e2ee_records (id, workspace_id, payload)
-                 VALUES (?, 'workspace-a', ?)",
-            )
-            .bind(&sealed.record_id)
-            .bind(&sealed.payload)
-            .execute(&mut *transaction)
-            .await
-            .unwrap();
-            sqlx::query(
-                "INSERT INTO e2ee_witness_records (
-                   workspace_id, record_id, revision, writer_id, payload_hash, payload, sequence
-                 ) VALUES (
-                   'workspace-a', ?, 1, 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa', ?, ?, ?
-                 )",
-            )
-            .bind(&sealed.record_id)
-            .bind(payload_hash)
-            .bind(&sealed.payload)
-            .bind(i64::from(index) + 1)
-            .execute(&mut *transaction)
-            .await
-            .unwrap();
-        }
-        transaction.commit().await.unwrap();
-
-        let checks = AtomicUsize::new(0);
-        let started_at = std::time::Instant::now();
-        let error = tokio::time::timeout(
-            std::time::Duration::from_secs(2),
-            apply_received_e2ee_replica_changes_with_witness_cancellable(
-                db.pool(),
-                &workspace_keys,
-                false,
-                || checks.fetch_add(1, AtomicOrdering::SeqCst) >= 20,
-            ),
-        )
-        .await
-        .expect("received-replica cancellation exceeded the activity deadline")
-        .unwrap_err();
-
-        assert!(matches!(error, E2eeReplicaError::Cancelled));
-        assert!(started_at.elapsed() < std::time::Duration::from_secs(2));
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>(
-                "SELECT COUNT(*) FROM sessions WHERE id LIKE 'remote-session-%'",
-            )
-            .fetch_one(db.pool())
-            .await
-            .unwrap(),
-            0
-        );
-        tokio::time::timeout(
-            std::time::Duration::from_millis(250),
-            sqlx::query(
-                "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-                 VALUES ('local-after-apply-cancel', 'workspace-a', 'user-a', 'Local')",
-            )
-            .execute(db.pool()),
-        )
-        .await
-        .expect("cancelled received-replica preflight kept the database busy")
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn bounded_received_preflight_skips_foreign_and_unwitnessed_prefixes() {
-        let workspace_keys = keys("workspace-a");
-        let key = &workspace_keys["workspace-a"];
-        let db = test_db().await;
-        sqlx::query(
-            "WITH RECURSIVE rows(id) AS (
-               SELECT 1
-               UNION ALL
-               SELECT id + 1 FROM rows WHERE id < 256
-             )
-             INSERT INTO e2ee_records (id, workspace_id, payload)
-             SELECT printf('!foreign-%04d', id), 'workspace-z', 'foreign' FROM rows",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "WITH RECURSIVE rows(id) AS (
-               SELECT 1
-               UNION ALL
-               SELECT id + 1 FROM rows WHERE id < 256
-             )
-             INSERT INTO e2ee_records (id, workspace_id, payload)
-             SELECT printf('!unwitnessed-%04d', id), 'workspace-a', 'unwitnessed' FROM rows",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        let sealed = key
-            .seal_field(
-                "workspace-a",
-                "sessions",
-                "witnessed-session",
-                ROW_MANIFEST_FIELD,
-                "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                1,
-                false,
-                json!(true),
-            )
-            .unwrap();
-        sqlx::query(
-            "INSERT INTO e2ee_records (id, workspace_id, payload)
-             VALUES (?, 'workspace-a', ?)",
-        )
-        .bind(&sealed.record_id)
-        .bind(&sealed.payload)
-        .execute(db.pool())
-        .await
-        .unwrap();
-        sqlx::query(
-            "INSERT INTO e2ee_witness_records (
-               workspace_id, record_id, revision, writer_id, payload_hash, payload, sequence
-             ) VALUES (
-               'workspace-a', ?, 1, 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa', ?, ?, 1
-             )",
-        )
-        .bind(&sealed.record_id)
-        .bind(anlg_e2ee::payload_hash(&sealed.payload))
-        .bind(&sealed.payload)
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let stats =
-            apply_received_e2ee_replica_changes_with_witness(db.pool(), &workspace_keys, false)
-                .await
-                .unwrap();
-
-        assert!(stats.rejected_unwitnessed >= 256);
-        assert_eq!(
-            sqlx::query_scalar::<_, i64>(
-                "SELECT COUNT(*) FROM sessions WHERE id = 'witnessed-session'",
-            )
-            .fetch_one(db.pool())
-            .await
-            .unwrap(),
-            1
-        );
-    }
-
-    #[tokio::test]
-    async fn id_collision_does_not_record_unmaterialized_remote_state() {
-        let recovery =
-            RecoveryKey::parse("anarlog-e2ee-v1:BwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwcHBwc")
-                .unwrap();
-        let workspace_keys = HashMap::from([
-            (
-                "workspace-a".to_string(),
-                recovery.workspace_key("workspace-a").unwrap(),
-            ),
-            (
-                "workspace-b".to_string(),
-                recovery.workspace_key("workspace-b").unwrap(),
-            ),
-        ]);
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-b', 'user-b', 'Remote B')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let target = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'Local A')",
-        )
-        .execute(target.pool())
-        .await
-        .unwrap();
-        copy_replica(source.pool(), target.pool()).await;
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let remote_state_count: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*) FROM e2ee_local_state
-             WHERE workspace_id = 'workspace-b'
-               AND table_name = 'sessions'
-               AND row_id = 'session-1'",
-        )
-        .fetch_one(target.pool())
-        .await
-        .unwrap();
-        let current_workspace: String =
-            sqlx::query_scalar("SELECT workspace_id FROM sessions WHERE id = 'session-1'")
-                .fetch_one(target.pool())
-                .await
-                .unwrap();
-        assert_eq!(remote_state_count, 0);
-        assert_eq!(current_workspace, "workspace-a");
-
-        sqlx::query("DELETE FROM sessions WHERE id = 'session-1'")
-            .execute(target.pool())
-            .await
-            .unwrap();
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let remote: (String, String) =
-            sqlx::query_as("SELECT workspace_id, title FROM sessions WHERE id = 'session-1'")
-                .fetch_one(target.pool())
-                .await
-                .unwrap();
-        assert_eq!(remote, ("workspace-b".to_string(), "Remote B".to_string()));
-    }
-
-    #[tokio::test]
-    async fn failed_remote_apply_rolls_back_the_guard() {
-        let workspace_keys = keys("workspace-a");
-        let source = test_db().await;
-        sqlx::query(
-            "INSERT INTO sessions (id, workspace_id, owner_user_id, title)
-             VALUES ('session-1', 'workspace-a', 'user-a', 'First')",
-        )
-        .execute(source.pool())
-        .await
-        .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-
-        let target = test_db().await;
-        copy_replica(source.pool(), target.pool()).await;
-        apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        sqlx::query("UPDATE sessions SET title = 'Remote' WHERE id = 'session-1'")
-            .execute(source.pool())
-            .await
-            .unwrap();
-        encrypt_e2ee_replica_changes(source.pool(), &workspace_keys)
-            .await
-            .unwrap();
-        let key = &workspace_keys["workspace-a"];
-        let title_record_id = key.blind_field_id("sessions", "session-1", "title");
-        let title_payload: String =
-            sqlx::query_scalar("SELECT payload FROM e2ee_records WHERE id = ?")
-                .bind(&title_record_id)
-                .fetch_one(source.pool())
-                .await
-                .unwrap();
-        sqlx::query("UPDATE e2ee_records SET payload = ? WHERE id = ?")
-            .bind(title_payload)
-            .bind(title_record_id)
-            .execute(target.pool())
-            .await
-            .unwrap();
-        sqlx::query(
-            "CREATE TRIGGER fail_remote_title_update
-             BEFORE UPDATE OF title ON sessions
-             BEGIN
-               SELECT RAISE(ABORT, 'forced apply failure');
-             END",
-        )
-        .execute(target.pool())
-        .await
-        .unwrap();
-
-        assert!(
-            apply_e2ee_replica_changes(target.pool(), &workspace_keys)
-                .await
-                .is_err()
-        );
-        let guard_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM e2ee_apply_guard")
-            .fetch_one(target.pool())
-            .await
-            .unwrap();
-        assert_eq!(guard_count, 0);
-
-        sqlx::query("DROP TRIGGER fail_remote_title_update")
-            .execute(target.pool())
-            .await
-            .unwrap();
-        sqlx::query("UPDATE sessions SET title = 'Local' WHERE id = 'session-1'")
-            .execute(target.pool())
-            .await
-            .unwrap();
-        let dirty_count: i64 = sqlx::query_scalar(
-            "SELECT COUNT(*) FROM e2ee_dirty_rows
-             WHERE workspace_id = 'workspace-a'
-               AND table_name = 'sessions'
-               AND row_id = 'session-1'",
-        )
-        .fetch_one(target.pool())
-        .await
-        .unwrap();
-        assert_eq!(dirty_count, 1);
-    }
-}
+mod tests;
